+++ public/index.html
@@ -1923,6 +1923,6 @@
       </section>
     </main>
 
-    <script src="../src/main.js" charset="utf-8"></script>
+    <script src="bundle.js" charset="utf-8"></script>
   </body>
 </html>
