+++ src/make-filter.js
@@ -1,0 +1,11 @@
+export default (filterName, taskCount) => `
+    <input
+      type="radio"
+      id="filter__${filterName}"
+      class="filter__input visually-hidden"
+      name="filter"
+      checked
+    />
+    <label for="filter__${filterName}" class="filter__label">
+      ${filterName} <span class="filter__${filterName}-count">${taskCount}</span></label>
+`;
